This is an automated PR to promote changes from `rc` to `master`.
Please review and test before merging.

See [TESTING.md](./TESTING.md) for complete testing instructions.

According to our release policy, this PR is expected to be merged on:
**1st or 3rd Friday of the month**
Testers are encouraged to test the changes before merging.
Please note that this schedule may be adjusted based on the needs of the
project.

---
* chore Merge branch 'master' into rc (28cd313f) by Rubin Bhandari
* fix (hyprlock): Potential fix for hyprlock crashing and env will now
use (5317adbb) by kRHYME7
* chore testing (45007cc7) by kRHYME7
* chore yeah (5e8778d4) by kRHYME7
* chore add manual ci (42dd1cfe) by kRHYME7
* chore ci:test rc-master (3f6dfa2e) by kRHYME7
* chore ci:test rc-master (9015a200) by kRHYME7
* chore Update README.md (3260be13) by Khing
* chore ci:test dev-rc (86897ba4) by kRHYME7
* chore chore: Added rc (Release Candidate) for the Frozen dev branch so
we can continue adding features in dev! (c58d4336) by kRHYME7
* fix: Correct Brave browser class name in windowrules.conf (#947)
(89ee3265) by GokuPlay609
* fix (waybar): https://github.com/prasanthrangan/hyprdots/issues/2154
(67a6df45) by kRHYME7
* fix > [#460
(comment)](https://github.com/HyDE-Project/HyDE/issues/460#issuecomment-2993943470)
#943 (46656c5e) by kRHYME7
* chore hesoyam (5b8b7e72) by kRHYME7
* chore chore: Try to extend testing time for our beloved testers.
(#937) (113deccb) by Khing
* chore refactor (waybar): Full cava module rewrite (faaceb73) by
kRHYME7
* fix: swaync (a47127e3) by kRHYME7
* chore refactor: hyde-shell is fork when executing script. (8e9b26cb)
by kRHYME7
* chore add best fzf aliases ever pt 2 And fish cleanups and bug fixes
(#919) (086ce030) by prime-run
* feat (hyprlock): greetd.conf with wallbash (70097fc0) by kRHYME7
* fix(hyprlock): greetd (8aae2b3f) by kRHYME7
* chore Add new hyprlock layout (#874) (4dcec19d) by Mahaveer Gurjar
* chore ci: try fix  warns on master (5daf073e) by kRHYME7
* chore Find symlinked themes (#924) (20d1dd7f) by Deepu K Sasidharan
* chore Use latest flags for vscode and electron (#885) (15794a26) by
Deepu K Sasidharan
* feat: Add 'hyde-shell version' to get local version (a7db67af) by
kRHYME7
* chore Make OMZ optional (#912) (e93a8ea7) by Deepu K Sasidharan
* chore remove -h (39576852) by kRHYME7
* chore added optional keybind "?" to append and run with --help.
(490f5003) by kRHYME7
* chore formatting terminal.zsh (237b92f5) by kRHYME7
* chore proper test operator for terminal.zsh (696ef3f1) by kRHYME7
* chore backport (shaders): Shaders v300 (#909) (83d35266) by Khing
* chore ok (c543ac71) by kRHYME7
* fix: add `prompt.zsh` to config restore list (#908) (adbfe43d) by Ilya
Bogdanov
* fix: do not use uwsm for hyprlock (#898) (7d143494) by Ilya Bogdanov
* chore refactor: chmod and some waybar fixes (24814f08) by kRHYME7
* feat: clean home directory (#867) (48c3184d) by Rubin Bhandari
* chore refactor: Porting our shaders to be #version 300 compatible
(#894) (c29bc661) by Khing
* chore Consistency, cleanup and minor improvements for waybar modules
(#883) (617dcbfd) by Deepu K Sasidharan
---

Please review the changes carefully before merging.

+++ Configs/.local/lib/hyde/workflows.sh
@@ -24,6 +24,7 @@ Usage: $0 [OPTIONS]
 
 Options:
     --select | -S       Select a workflow from the available options
+    --set               Set the given workflow
     --waybar            Get workflow info for Waybar
     --help   | -h       Show this help message
     
@@ -36,7 +37,7 @@ if [ -z "${*}" ]; then
 fi
 
 # Define long options
-LONG_OPTS="select,waybar,help"
+LONG_OPTS="select,set:,waybar,help"
 SHORT_OPTS="Sh"
 # Parse options
 PARSED=$(getopt --options ${SHORT_OPTS} --longoptions "${LONG_OPTS}" --name "$0" -- "$@")
@@ -67,7 +68,7 @@ fn_select() {
     workflow_icon=$(get_hyprConf "WORKFLOW_ICON" "$workflow_path")
     workflow_icon=${workflow_icon:0:1}
     workflow_list="${workflow_list}\n${workflow_icon}\t ${workflow_name}"
-  done < <(find "$workflows_dir" -type f -name "*.conf" 2>/dev/null)
+  done < <(find "$workflows_dir" -follow -type f -name "*.conf" 2>/dev/null)
 
   # Set rofi scaling
   font_scale="${ROFI_WORKFLOW_SCALE}"
@@ -144,7 +145,7 @@ fn_update() {
 #*│                                                                            │
 #*└────────────────────────────────────────────────────────────────────────────┘
 
-\$WORKFLOW = ${ucrrent_workflow}
+\$WORKFLOW = ${current_workflow}
 \$WORKFLOW_ICON = ${current_icon}
 \$WORKFLOW_DESCRIPTION = ${current_description}
 \$WORKFLOWS_PATH = ./workflows/${current_workflow}.conf
@@ -172,6 +173,20 @@ while true; do
   case "$1" in
   -S | --select)
     fn_select
+    # refresh waybar module only if waybar is running
+    if pgrep -x waybar >/dev/null; then
+      pkill -RTMIN+7 waybar
+    fi
+
+    exit 0
+    ;;
+  --set)
+    if [ -z "$2" ]; then
+      echo "Error: --set requires a workflow name"
+      exit 1
+    fi
+    set_conf "HYPR_WORKFLOW" "$2"
+    fn_update
     # refresh waybar module only if waybar is running
     if pgrep -x waybar >/dev/null; then
       pkill -RTMIN+7 waybar
